+++ lib/generators/feature_generator.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:path/path.dart' as p;
 import 'package:rekeens_flutter_cli/generators/base_generator.dart';
 import 'package:rekeens_flutter_cli/services/router_updater.dart';
 import 'package:rekeens_flutter_cli/utils/logger.dart';
@@ -19,6 +20,7 @@ class FeatureGenerator extends BaseGenerator {
     String featureName, {
     bool force = false,
     bool dryRun = false,
+    bool withTests = true,
   }) async {
     if (featureName.isEmpty) {
       throw Exception('Feature name is required.');
@@ -36,8 +38,21 @@ class FeatureGenerator extends BaseGenerator {
 
     final className = toPascalCase(featureName);
 
+    final testDir = p.join(
+      projectDir,
+      'test',
+      'features',
+      featureName,
+      'presentation',
+      'pages',
+    );
+    final testPath = p.join(testDir, '${featureName}_page_test.dart');
+
     if (dryRun) {
       logDryRun('create feature "$featureName"', featureDir);
+      if (withTests) {
+        logDryRun('create feature page test "$featureName"', testPath);
+      }
       await _routerUpdater.addRoute(
         featureName: featureName,
         className: '${className}Page',
@@ -57,6 +72,18 @@ class FeatureGenerator extends BaseGenerator {
       targetDir: featureDir,
       variables: variables,
     );
+
+    if (withTests) {
+      await generateTest(
+        testTemplateSubPath: 'feature',
+        testDir: testDir,
+        variables: {
+          'project_name': getProjectName(),
+          'feature_name': featureName,
+          'class_name': className,
+        },
+      );
+    }
 
     await _routerUpdater.addRoute(
       featureName: featureName,